// frontend/src/app/core/api/model/room-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RoomResponse { 
    id?: number;
    code?: string;
    name?: string;
    emplacement?: string;
    capacity?: number;
    type?: RoomResponse.TypeEnum;
    status?: RoomResponse.StatusEnum;
}
export namespace RoomResponse {
    export const TypeEnum = {
        Classroom: 'CLASSROOM',
        ComputerLab: 'COMPUTER_LAB',
        Workshop: 'WORKSHOP',
        MeetingRoom: 'MEETING_ROOM'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
    export const StatusEnum = {
        Draft: 'DRAFT',
        Active: 'ACTIVE',
        Maintenance: 'MAINTENANCE',
        Archived: 'ARCHIVED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}