// frontend/src/app/core/api/model/room-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RoomRequest { 
    name: string;
    capacity: number;
    type: RoomRequest.TypeEnum;
    emplacement: string;
}
export namespace RoomRequest {
    export const TypeEnum = {
        Classroom: 'CLASSROOM',
        ComputerLab: 'COMPUTER_LAB',
        Workshop: 'WORKSHOP',
        MeetingRoom: 'MEETING_ROOM'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}